[CA-220467] CID-28703 Destination buffer too small
Added check to make sure destination buffer is big enaugh
for the name wich is going to be copied in.

+++ drivers/tapdisk-nbdserver.c
@@ -834,6 +834,14 @@ tapdisk_nbdserver_listen_unix(td_nbdserver_t *server)
 	}
 
 	server->local.sun_family = AF_UNIX;
+
+	if (unlikely(strlen(server->sockpath) > 
+		     (sizeof(server->local.sun_path) - 1))) {
+		err = -ENAMETOOLONG;
+		ERR("socket name too long: %s\n", server->sockpath);
+		goto out;
+	}
+
 	strcpy(server->local.sun_path, server->sockpath);
 	err = unlink(server->local.sun_path);
 	if (err == -1 && errno != ENOENT) {
